refactor: ENUM 변환 예외 처리 공용 UnsupportedException로 변경 #11
