--- src/main/java/com/dh/ondot/schedule/domain/Mission.java
@@ -1,4 +1,0 @@
-package com.dh.ondot.schedule.domain;
-
-public enum Mission {
-}
